src/requirements.txt
python-telegram-bot
grpcio
grpcio-tools
pymongo
aiocryptopay
matplotlib
pandas
adjustText
reportlab
